Working version backup.

+++ Fossil/src/GameObject.h
@@ -5,6 +5,7 @@
 #include "BuildDll.h"
 #include "Settings.h"
 #include "glew.h"
+#include "glfw3.h"
 
 class FSL_API GameObject
 {
@@ -14,7 +15,7 @@ public:
 	std::vector<GLfloat> vertices;
 	std::vector<GLfloat> textures;
 	std::vector<GLfloat> normals;
-	std::vector<int> faces;
+	std::vector<GLushort> faces;
 	GLfloat xpos;
 	GLfloat ypos;
 	GLfloat zpos;
@@ -24,7 +25,7 @@ public:
 	GameObject(std::string name, std::string file, GLfloat x, GLfloat y, GLfloat z);
 	~GameObject();
 	void ImportOBJ();
-	void ConvertToArrays(GLfloat* verts, GLfloat* txtrs, GLfloat* norms, int* face);
+	void ConvertToArrays(GLfloat* verts, GLfloat* txtrs, GLfloat* norms, GLushort* face);
 	void DrawObject();
 
 	void MoveObject(GLfloat xdist, GLfloat ydist, GLfloat zdist);
